requirements.txt
ecdsa
pycryptodome
requests
socket
json
hashlib
os